fixed warning on AMD64
git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@15314 9d14d920-75d5-0310-8bd7-a52572981aee

+++ engine/builder/coverage/tce_dump.c
@@ -47,7 +47,7 @@ main(int argc, char *argv[])
     int peer_id;
 
     static char ta_list[4096];
-    int ta_list_len = sizeof(ta_list);
+    size_t ta_list_len = sizeof(ta_list);
     char *ta;
 
     char buffer[PATH_MAX + 1];
